backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Database - Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
itsdangerous==2.1.2

# OAuth
httpx==0.26.0

# Anthropic Claude API
anthropic==0.18.1

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Email
sendgrid==6.11.0

# Background Tasks
celery==5.3.6
flower==2.0.1

# WebSocket
python-socketio==5.11.1
python-socketio[asyncio-client]==5.11.1

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Environment Variables
python-dotenv==1.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Requests
aiohttp==3.9.3
requests==2.31.0

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# Rate Limiting
slowapi==0.1.9

# Logging & Monitoring
sentry-sdk==1.40.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.6.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
pytz==2024.1
python-slugify==8.0.2
